Revert "refactored server"
