--- modules/notFound.js
@@ -1,8 +1,0 @@
-'use srtict';
-
-function notFound(request, response) {
-  response.status(404).send('Not Found');
-
-}
-
-module.exports = notFound;
